// frontend/artifact-depot-app/src/app/types/artifact.ts
export interface Artifact{
    artifactId?: number;
    artifactName: string;
    width: number;
    height: number;
    artifactLatitude:number;
    artifactLongitude:number;  
    isDeleted: boolean;
    artifactMaterialId: number;
    artifactCategoryId: number;
    excavationSiteId: number;
    
  }
  